Movement on a hex-grid system is done. read description for more.
I also added a end turn function, not sure if that counts as a base turn or not. I moved all the other stuff I had previously into a new folder called _OLD. As far I know, the movement script will work on a hexagonal grid system :) Just may have trouble moving it over and implementing it, but figured we'll cross that bridge when we get to it. I'm too mentally broken at the moment and its 3.33am